+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeatureTest.kt
@@ -80,6 +80,10 @@ class ToolbarAutocompleteFeatureTest {
         override fun editMode() {
             fail()
         }
+
+        override fun addEditAction(action: Toolbar.Action) {
+            fail()
+        }
     }
 
     @Test